Reorganize plugin API as statscache.plugins
Relocate all modules for use by plugins under the statscache.plugins subpackage.
The BasePlugin class is placed in the __init__.py of the subpackage, and the
model classes are split off into statscache.plugins.models. All classes
previously available in the statscache.plugins module is re-exported by the
__init__.py of the package as well as the Schedule class, now implemented in
statscache.plugins.schedule.
