--- statscache/api/__init__.py
@@ -1,4 +1,0 @@
-from statscache.api.plugins import BaseModel, ScalarModel,\
-                                   CategorizedModel, CategorizedLogModel,\
-                                   ConstrainedCategorizedLogModel
-from statscache.api.schedule import Schedule
