+++ statscache/plugins/schedule.py
